user authentication package
Models user as a NamedResource with and ssh/rsa public key and a
specific role.

A role is defined as a set of privileges and a privelege captures the
create, read, update and deleted operations against a resource key. If
the resource key is empty then the privilege applies to all resources.

Signed-off-by: Ravi Chamarthy <ravi@chamarthy.dev>