work so far, still very unfinished
threw some ideas down for lease, added some functions
very buggy -- does not work as intended

Basically, lease is managed through ResourceAPI (map of group pointers which each are associated with a lease.Queue pointer).
When store is initialized, we also initialize the lease queues. To do so, we first need to create groups.
	- Do this by going through resources from store and create groups as we find new labels. Add resources to their group and assign group a queue.

When we have our api.LeaseQueue initialized, we can now add a lease to all the group queues for groups present in the request.
After this, we process each group queue individually until all groups have completed. Then the lease is ready for activation.

I have not yet implemented this parallel group queue processing and the lease is considered "processed" after assigning whatever free resources
are available. This isn't correct and will be fixed in next commits. Additionally, have to clean up code to be properly organized.

Signed-off-by: Shravya Nandyala <shravya.nandyala@gmail.com>